Refactor: Add conftest.py and dummy.py
