+++ tests/test_devices.py
@@ -1,126 +1,27 @@
-from collections.abc import Generator
 import math
 from pathlib import Path
-import sys
 
 from samplemaker.baselib.waveguides import BaseWaveguidePort, BaseWaveguideSequencer
 import samplemaker.devices as smdev
 import samplemaker.shapes as sp
 import pytest
-from fixtures import reset_samplemaker  # noqa: F401
 
-from samplemaker.makers import make_rect
 from samplemaker.shapes import GeomGroup
-
-
-class DummyDevice(smdev.Device):
-    """Simple test device.
-
-    Consists of a rectangle with a single port on the left side.
-    The width and height of the rectangle can be set via parameters.
-    """
-
-    def initialize(self):
-        self.set_name("TESTLIB_DUMMY")
-        self.set_description("A dummy device for testing purposes.")
-
-    def parameters(self):
-        self.addparameter(
-            param_name="width",
-            default_value=10.0,
-            param_description="Width of the sample.",
-            param_type=float,
-            param_range=(1.0, 1000.0),
-        )
-        self.addparameter(
-            param_name="height",
-            default_value=5.0,
-            param_description="Height of the sample.",
-            param_type=float,
-            param_range=(1.0, 1000.0),
-        )
-
-    def geom(self):
-        p = self.get_params()
-        width = p["width"]
-        height = p["height"]
-
-        rectangle = make_rect(0, 0, width, height, numkey=4)
-        port = BaseWaveguidePort(0, 0, orient="West", name="in")
-        self.addlocalport(port)
-
-        return rectangle
-
-
-def _dummy_connector(_port1: smdev.DevicePort, _port2: smdev.DevicePort) -> GeomGroup:
-    return GeomGroup()
-
-
-class ConnectorPort(smdev.DevicePort):
-    def __init__(
-        self, x0: float, y0: float, horizontal: bool, forward: bool, name: str
-    ):
-        super().__init__(x0, y0, horizontal, forward)
-        self.name = name
-        self.connector_function = _dummy_connector
-
-
-class ConnectorDevice(smdev.Device):
-    def initialize(self):
-        self.set_name("TESTLIB_CONNECTOR")
-        self.set_description("Simple connector-compatible test device")
-
-    def parameters(self):
-        self.addparameter(
-            param_name="length",
-            default_value=10.0,
-            param_description="Length of test geometry",
-            param_type=float,
-        )
-
-    def geom(self):
-        p = self.get_params()
-        rect = make_rect(0, 0, p["length"], 1, numkey=4)
-        self.addlocalport(ConnectorPort(0, 0, True, True, name="io"))
-        return rect
-
-
-@pytest.fixture
-def default_device_list() -> Generator[dict[str, type[smdev.Device]], None, None]:
-    old_devlist = smdev._DeviceList.copy()
-    yield smdev._DeviceList
-    smdev._DeviceList.clear()
-    smdev._DeviceList.update(old_devlist)
-
-
-@pytest.fixture
-def device_list(
-    default_device_list: dict[str, type[smdev.Device]],
-) -> Generator[dict[str, type[smdev.Device]], None, None]:
-    dev = DummyDevice()
-    dev.initialize()
-    default_device_list[dev._name] = DummyDevice
-    return default_device_list
-
-
-@pytest.fixture
-def connector_device_list(
-    default_device_list: dict[str, type[smdev.Device]],
-) -> Generator[dict[str, type[smdev.Device]], None, None]:
-    dev = ConnectorDevice()
-    dev.initialize()
-    default_device_list[dev._name] = ConnectorDevice
-    return default_device_list
+import dummy as dm
 
 
 @pytest.fixture
 def simple_netlist(
-    connector_device_list: dict[str, type[smdev.Device]],
+    dummy_device_list: dict[str, type[smdev.Device]],
 ) -> smdev.NetList:
-    _ = connector_device_list
+    _ = dummy_device_list
     netlist_entries = [
-        smdev.NetListEntry("TESTLIB_CONNECTOR", 0.0, 0.0, "E", {"io": "wire1"}, {}),
-        smdev.NetListEntry("TESTLIB_CONNECTOR", 20.0, 0.0, "W", {"io": "wire1"}, {}),
+        smdev.NetListEntry(
+            "TESTLIB_DUMMY_CONNECTOR", 0.0, 0.0, "E", {"io": "wire1"}, {}
+        ),
+        smdev.NetListEntry(
+            "TESTLIB_DUMMY_CONNECTOR", 20.0, 0.0, "W", {"io": "wire1"}, {}
+        ),
     ]
     return smdev.NetList("simple", netlist_entries)
 
@@ -237,7 +138,7 @@ class TestDevicePort:
 
 class TestDevice:
     def test_initialize_sets_default_attributes(self) -> None:
-        dev = DummyDevice()
+        dev = dm.DummyDevice()
         assert dev._p == {}
         assert dev._pdescr == {"_ports_": "Ports calculated by geom"}
         assert "_ports_" in dev._ptype  # Is really float...
@@ -253,17 +154,17 @@ class TestDevice:
         assert dev.use_references is True
 
     def test_build_initializes_device(
-        self, device_list: dict[str, type[smdev.Device]]
+        self, dummy_device_list: dict[str, type[smdev.Device]]
     ) -> None:
-        _ = device_list
-        dev = DummyDevice.build()
+        _ = dummy_device_list
+        dev = dm.DummyDevice.build()
 
         assert dev._name == "TESTLIB_DUMMY"
         assert dev._description == "A dummy device for testing purposes."
         assert set(dev._p.keys()) == {"width", "height"}
 
     def test_name_and_params_affect_hash(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         h1 = hash(dev)
         dev.set_param("width", 20.0)
         h2 = hash(dev)
@@ -272,7 +173,7 @@ class TestDevice:
         assert h1 != h2 != h3
 
     def test_sequencer_options_affect_hash(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         h1 = hash(dev)
         dev._seq = BaseWaveguideSequencer([])
         h2 = hash(dev)
@@ -292,14 +193,14 @@ class TestDevice:
     def test_get_set_angle(
         self, angle: float, expected_hv: bool, expected_bf: bool
     ) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         dev.set_angle(angle)
         dev._hv = expected_hv
         dev._bf = expected_bf
         assert dev.angle() == pytest.approx(angle)
 
     def test_set_position(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         dev.set_position(10.0, 20.0)
         assert dev._x0 == pytest.approx(10.0)
         assert dev._y0 == pytest.approx(20.0)
@@ -310,22 +211,22 @@ class TestDevice:
         assert bb.cy() == pytest.approx(20.0)
 
     def test_addparameter_rejects_colon(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         with pytest.raises(ValueError, match="containing ':'"):
             dev.addparameter("bad:name", 1.0, "invalid")
 
     def test_addlocalparameter_rejects_colon(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         with pytest.raises(ValueError, match="containing ':'"):
             dev.addlocalparameter("bad:name", 1.0, "invalid")
 
     def test_set_param_unknown_raises(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         with pytest.raises(ValueError, match="Could not set parameter"):
             dev.set_param("does_not_exist", 1)
 
     def test_get_params_casts_and_clips(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         dev.set_param("width", "1234")
         dev.set_param("height", -23)
         p = dev.get_params(cast_types=True, clip_in_range=True)
@@ -343,12 +244,12 @@ class TestDevice:
             p = dev.get_params(cast_types=True, clip_in_range=True)
 
     def test_get_localport_raises_for_missing_port(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         with pytest.raises(ValueError, match="Could not find port"):
             dev.get_localport("missing")
 
     def test_remove_localport(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         port = BaseWaveguidePort(0, 0, name="test")
         dev.addlocalport(port)
         assert "test" in dev._localp["_ports_"]
@@ -360,21 +261,21 @@ class TestDevice:
         reason="remove_localport should raise for missing port but currently does not",
     )
     def test_remove_localport_unknown_raises(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         with pytest.raises(ValueError, match="Could not find local port"):
             dev.remove_localport("missing")
 
     def test_get_port_raises_for_missing_port(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         with pytest.raises(ValueError, match="Could not find port"):
             dev.get_port("missing")
 
     def test_build_registered_returns_named_device(
-        self, device_list: dict[str, type[smdev.Device]]
+        self, dummy_device_list: dict[str, type[smdev.Device]]
     ) -> None:
-        _ = device_list
+        _ = dummy_device_list
         dev = smdev.Device.build_registered("TESTLIB_DUMMY")
-        assert isinstance(dev, DummyDevice)
+        assert isinstance(dev, dm.DummyDevice)
 
     def test_build_registered_unknown_name_raises(
         self, default_device_list: dict[str, type[smdev.Device]]
@@ -384,7 +285,7 @@ class TestDevice:
             smdev.Device.build_registered("UNKNOWN")
 
     def test_run_use_reference(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         g = dev.run()
         assert "in" in dev._ports
 
@@ -404,7 +305,7 @@ class TestDevice:
         assert isinstance(g.group[0], sp.SRef)
 
     def test_run_no_reference(self) -> None:
-        dev = DummyDevice.build()
+        dev = dm.DummyDevice.build()
         dev.use_references = False
         g = dev.run()
         assert "in" in dev._ports
@@ -453,7 +354,7 @@ class TestNetList:
         circuit_file = tmp_path / "test_circuit.txt"
         lines = [
             ".CIRCUIT CHILD out",
-            "TESTLIB_CONNECTOR 0 0 E io out .",
+            "TESTLIB_DUMMY_CONNECTOR 0 0 E io out .",
             ".END",
             ".CIRCUIT TOP ext",
             "X CHILD 10 0 E io ext .",
@@ -471,7 +372,7 @@ class TestNetList:
         circuit_file = tmp_path / "single_circuit.txt"
         lines = [
             ".CIRCUIT MAIN wire",
-            "TESTLIB_CONNECTOR 0 0 E io wire .",
+            "TESTLIB_DUMMY_CONNECTOR 0 0 E io wire .",
             ".END",
         ]
         circuit_file.write_text("\n".join(lines))
@@ -488,7 +389,7 @@ class TestNetList:
             ".CIRCUIT MAIN ext",
             ".ALIGN wire1 wire2",
             ".PATH wire1 0 0 E 5 5 N 10 5 W",
-            "TESTLIB_CONNECTOR 0 0 E io wire1 .",
+            "TESTLIB_DUMMY_CONNECTOR 0 0 E io wire1 .",
             ".END",
         ]
         circuit_file.write_text("\n".join(lines))
@@ -502,22 +403,22 @@ class TestNetList:
 class TestCircuit:
     def test_set_param_netlist_updates_hierarchical_parameters(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         simple_netlist: smdev.NetList,
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         circuit = smdev.Circuit.build()
         circuit.set_param("NETLIST", simple_netlist)
 
-        assert "dev_TESTLIB_CONNECTOR_1" in circuit._p
-        assert "dev_TESTLIB_CONNECTOR_2" in circuit._p
+        assert "dev_TESTLIB_DUMMY_CONNECTOR_1" in circuit._p
+        assert "dev_TESTLIB_DUMMY_CONNECTOR_2" in circuit._p
 
     def test_run_connects_two_compatible_ports(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         simple_netlist: smdev.NetList,
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         circuit = smdev.Circuit.build()
         circuit.set_param("NETLIST", simple_netlist)
 
@@ -526,11 +427,13 @@ class TestCircuit:
 
     def test_run_warns_for_unconnected_non_external_port(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         netlist_entries = [
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 0, 0, "E", {"io": "dangling"}, {})
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 0, 0, "E", {"io": "dangling"}, {}
+            )
         ]
         netlist = smdev.NetList("warn", netlist_entries)
         circuit = smdev.Circuit.build()
@@ -541,11 +444,13 @@ class TestCircuit:
 
     def test_run_exports_external_ports(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         netlist_entries = [
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 0, 0, "E", {"io": "wire_out"}, {})
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 0, 0, "E", {"io": "wire_out"}, {}
+            )
         ]
         netlist = smdev.NetList("ext", netlist_entries)
         netlist.set_external_ports(["wire_out"])
@@ -558,10 +463,10 @@ class TestCircuit:
 
     def test_run_aligns_ports_on_y_when_aligned_wire_exists(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         captured: list[tuple[float, float, float, float]] = []
 
         def _capture_connector(
@@ -570,12 +475,14 @@ class TestCircuit:
             captured.append((port1.x0, port1.y0, port2.x0, port2.y0))
             return GeomGroup()
 
-        monkeypatch.setattr(
-            sys.modules[__name__], "_dummy_connector", _capture_connector
-        )
+        monkeypatch.setattr(dm, "_dummy_connector", _capture_connector)
         netlist_entries = [
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 0.0, 0.0, "E", {"io": "w1"}, {}),
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 20.0, 5.0, "W", {"io": "w1"}, {}),
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 0.0, 0.0, "E", {"io": "w1"}, {}
+            ),
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 20.0, 5.0, "W", {"io": "w1"}, {}
+            ),
         ]
         netlist = smdev.NetList("align_y", netlist_entries)
         netlist.set_aligned_ports(["w1"])
@@ -589,10 +496,10 @@ class TestCircuit:
 
     def test_run_aligns_ports_on_x_for_vertical_ports(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         captured: list[tuple[float, float, float, float]] = []
 
         def _capture_connector(
@@ -601,13 +508,15 @@ class TestCircuit:
             captured.append((port1.x0, port1.y0, port2.x0, port2.y0))
             return GeomGroup()
 
-        monkeypatch.setattr(
-            sys.modules[__name__], "_dummy_connector", _capture_connector
-        )
+        monkeypatch.setattr(dm, "_dummy_connector", _capture_connector)
 
         netlist_entries = [
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 0.0, 0.0, "N", {"io": "w1"}, {}),
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 5.0, 20.0, "S", {"io": "w1"}, {}),
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 0.0, 0.0, "N", {"io": "w1"}, {}
+            ),
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 5.0, 20.0, "S", {"io": "w1"}, {}
+            ),
         ]
         netlist = smdev.NetList("align_x", netlist_entries)
         netlist.set_aligned_ports(["w1"])
@@ -621,10 +530,10 @@ class TestCircuit:
 
     def test_run_routes_connector_through_virtual_path_points(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         captured: list[tuple[float, float, float, float]] = []
 
         def _capture_connector(
@@ -633,13 +542,15 @@ class TestCircuit:
             captured.append((port1.x0, port1.y0, port2.x0, port2.y0))
             return GeomGroup()
 
-        monkeypatch.setattr(
-            sys.modules[__name__], "_dummy_connector", _capture_connector
-        )
+        monkeypatch.setattr(dm, "_dummy_connector", _capture_connector)
 
         netlist_entries = [
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 0.0, 0.0, "E", {"io": "w1"}, {}),
-            smdev.NetListEntry("TESTLIB_CONNECTOR", 20.0, 0.0, "W", {"io": "w1"}, {}),
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 0.0, 0.0, "E", {"io": "w1"}, {}
+            ),
+            smdev.NetListEntry(
+                "TESTLIB_DUMMY_CONNECTOR", 20.0, 0.0, "W", {"io": "w1"}, {}
+            ),
         ]
         netlist = smdev.NetList("path_route", netlist_entries)
         netlist.set_path("w1", [10.0, 0.0, 0.0])
@@ -659,11 +570,11 @@ class TestCircuit:
 class TestDeviceLibraryExports:
     def test_create_device_library(
         self,
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         tmp_path: Path,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
-        _ = connector_device_list
+        _ = dummy_device_list
         output_file = tmp_path / "library.gds"
 
         calls: dict[str, object] = {}
@@ -682,13 +593,13 @@ class TestDeviceLibraryExports:
         monkeypatch.setattr(smdev, "GDSWriter", FakeGDSWriter)
 
         smdev.CreateDeviceLibrary(
-            "TESTLIB_CONNECTOR", {"length": 7.0}, str(output_file)
+            "TESTLIB_DUMMY_CONNECTOR", {"length": 7.0}, str(output_file)
         )
 
         geom = calls["geom"]
 
         assert calls["open"] == str(output_file)
-        assert calls["devname"] == "TESTLIB_CONNECTOR"
+        assert calls["devname"] == "TESTLIB_DUMMY_CONNECTOR"
         assert calls["closed"] is True
         assert isinstance(geom, GeomGroup)
         assert isinstance(geom.group[-1], sp.Text)
@@ -701,22 +612,21 @@ class TestDeviceLibraryExports:
 
     def test_export_device_schematics(
         self,
-        device_list: dict[str, type[smdev.Device]],
-        connector_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
         tmp_path: Path,
     ) -> None:
-        _ = (device_list, connector_device_list)
+        _ = dummy_device_list
         output_file = tmp_path / "SampleMakerLibrary.lel"
 
         smdev.ExportDeviceSchematics(str(output_file))
         content = output_file.read_text()
 
         assert "<Component DummyDevice>" in content
-        assert "<Component ConnectorDevice>" in content
+        assert "<Component DummyConnectorDevice>" in content
         assert "<Description>" in content
         assert "<Parameter>" in content
         assert "<Prefix TESTLIB_DUMMY>" in content
-        assert "<Prefix TESTLIB_CONNECTOR>" in content
+        assert "<Prefix TESTLIB_DUMMY_CONNECTOR>" in content
         assert "<Netlist spice>" in content
         assert "<Prefix X>" not in content
 
@@ -724,9 +634,9 @@ class TestDeviceLibraryExports:
 class TestDeviceRegistration:
     def test_register_devices_in_module(
         self,
-        default_device_list: dict[str, type[smdev.Device]],
+        dummy_device_list: dict[str, type[smdev.Device]],
     ) -> None:
-        _ = default_device_list
+        _ = dummy_device_list
         smdev.registerDevicesInModule(__name__)
         assert "TESTLIB_DUMMY" in smdev._DeviceList
-        assert "TESTLIB_CONNECTOR" in smdev._DeviceList
+        assert "TESTLIB_DUMMY_CONNECTOR" in smdev._DeviceList
